requirements.txt
graia-application
graia-broadcast
graia-scheduler
pymysql